develop
I'm kind of in a torn state as to whether or not to fully develop the client OR just finish cache loading for blueprints. I will start with just the cache loading.

This loads a cache located in %USER%/rs2-225-cs/

As of this commit, It is capable of decoding FloType Configs.